environments/alphazero_llm_trainer/pyproject.toml
[project]
name = "alphazero-llm-trainer"
description = "AlphaZero-inspired MCTS environment for training LLMs through tree search and policy learning with teacher ensemble guidance and dual reward systems"
keywords = ["placeholder-tag", "train", "eval"]
version = "0.1.2"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.2.0",
    "numpy>=2.0.0,<2.3",  # Changed: numba 0.61.2 requires numpy<=2.2
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
    "pyyaml",
    "scikit-learn>=1.7.2",
    "torch==2.8.0",
    "unsloth>=2024.8",
    "verifiers>=0.1.5.post0",
    "vllm>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "core", "models", "prompts", "rewards", "utils"]
